requirements.txt
fastapi-poe>=0.0.46
pyswisseph>=2.10.0
timezonefinder>=6.2.0
geopy>=2.4.0
